requirements.txt
# Core
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM providers
google-genai>=1.0.0
groq>=0.4.0
langgraph>=0.0.30

# Data
pandas>=2.0.0
requests>=2.31.0
